starting to implement collisions
collision script and the physics script act like the visual script and the rendering script:
one contains data types and the other the implementation of it.
collisions are based on an integer point array that form shapes easily adaptable to the low res screen and its sprites.

******BUG******
the (is object colliding with other object) func is not working as of now.
some parts of the code have been tested and maked like so using (# tested)